+++ Dyna/ExternalAPIs/IAppleShipyardApi.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Shockah.Dyna;
+
+public interface IAppleShipyardApi
+{
+	void RegisterActionLooksForPartType(Type actionType, PType partType);
+}